refactor(server): keep only one controller for User business model
