+++ packages/server/start/routes.ts
@@ -45,7 +45,7 @@ Route.group(() => {
     Route.post('/search/followers', 'SearchUsersController.listUserFollowers');
     Route.post('/search/following', 'SearchUsersController.listUserFollowing');
 
-    Route.post('/search/mpe', 'UsersController.listUserMpeRooms');
+    Route.post('/search/mpe', 'UserProfileController.listUserMpeRooms');
 }).prefix(USER_ROUTES_GROUP_PREFIX);
 
 export const AUTHENTICATION_ROUTES_GROUP_PREFIX = '/authentication';
